stop using mem::forget
We want to prevent some temporary Files from closing the fd they contain
when they drop, and mem::forget is one way to do that, but it assumes
that files don't have any heap memory attached to them that needs to be
freed. That's true right now, but it's better not to assume it'll always
be true. IntoRawFd/IntoRawHandle should take care of freeing memory, if
there ever is any.

Also, add a test that fails if we forget this no-drop logic entirely.