fix a bug where cloned pipes were inheritable on Windows
There is a bug (I think it's a bug?) in libstd which makes
File::try_clone return inheritable handles. We need to work around that
by making the same system calls explictly, but with the inheritable flag
set to false. I've filed a fix against libstd
(https://github.com/rust-lang/rust/pull/65316), but even if that lands
we'll need to retain this workaround for all the past versions of Rust
that have the bug.